Reinclude configs and contexts + cleanup
