--- decapodes/askemdecapodes_schema.json
@@ -1,65 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "$defs": {
-    "ASKEMDeca": {
-      "oneOf": [
-        {
-          "$ref": "#/$defs/ASKEMDecaExpr"
-        },
-        {
-          "$ref": "#/$defs/ASKEMDecapode"
-        }
-      ]
-    },
-    "ASKEMDecaExpr": {
-      "type": "object",
-      "properties": {
-        "tag": {
-          "const": "ASKEMDecaExpr"
-        },
-        "header": {
-          "$ref": "#/$defs/amr.Header"
-        },
-        "model": {
-          "$ref": "#/$defs/decapodes.DecaExpr"
-        },
-        "annotations": {
-          "type": "array",
-          "items": {
-            "$ref": "#/$defs/amr.Annotation"
-          }
-        }
-      },
-      "required": [
-        "header",
-        "model",
-        "annotations"
-      ]
-    },
-    "ASKEMDecapode": {
-      "type": "object",
-      "properties": {
-        "tag": {
-          "const": "ASKEMDecapode"
-        },
-        "header": {
-          "$ref": "#/$defs/amr.Header"
-        },
-        "model": {
-          "$ref": "#/$defs/decapodeacset.SymSummationDecapode"
-        },
-        "annotations": {
-          "type": "array",
-          "items": {
-            "$ref": "#/$defs/amr.Annotation"
-          }
-        }
-      },
-      "required": [
-        "header",
-        "model",
-        "annotations"
-      ]
-    }
-  }
-}
