+++ decapodes/json/configuration_schema.json
@@ -1,0 +1,192 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$defs": {
+    "Parameter": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/LiteralParameter"
+        },
+        {
+          "$ref": "#/$defs/ConstantParameter"
+        }
+      ]
+    },
+    "LiteralParameter": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "LiteralParameter"
+        },
+        "type": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "value": {
+          "type": "number",
+          "$comment": "F64"
+        }
+      },
+      "required": [
+        "type",
+        "value"
+      ]
+    },
+    "ConstantParameter": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "ConstantParameter"
+        },
+        "type": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "value": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "type",
+        "value"
+      ]
+    },
+    "Condition": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "value": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "domain_mesh": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "type",
+        "value",
+        "domain_mesh"
+      ]
+    },
+    "DatasetFormat": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/NetCDF"
+        },
+        {
+          "$ref": "#/$defs/CSV"
+        }
+      ]
+    },
+    "NetCDF": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "NetCDF"
+        }
+      },
+      "required": [
+      ]
+    },
+    "CSV": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "CSV"
+        }
+      },
+      "required": [
+      ]
+    },
+    "DatasetFile": {
+      "type": "object",
+      "properties": {
+        "uri": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "format": {
+          "$ref": "#/$defs/DatasetFormat"
+        },
+        "shape": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "$comment": "I64"
+          }
+        }
+      },
+      "required": [
+        "uri",
+        "format",
+        "shape"
+      ]
+    },
+    "Dataset": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "name": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "description": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "file": {
+          "$ref": "#/$defs/DatasetFile"
+        }
+      },
+      "required": [
+        "type",
+        "name",
+        "description",
+        "file"
+      ]
+    },
+    "DecapodeConfiguration": {
+      "type": "object",
+      "properties": {
+        "parameters": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/Parameter"
+          }
+        },
+        "initial_conditions": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/Condition"
+          }
+        },
+        "boundary_conditions": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/Condition"
+          }
+        },
+        "datasets": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/Dataset"
+          }
+        }
+      },
+      "required": [
+        "parameters",
+        "initial_conditions",
+        "boundary_conditions",
+        "datasets"
+      ]
+    }
+  }
+}
