+++ decapodes/json/context_schema.json
@@ -1,0 +1,343 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$defs": {
+    "File": {
+      "type": "object",
+      "properties": {
+        "uri": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "format": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "uri",
+        "format"
+      ]
+    },
+    "Constant": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/FileConstant"
+        },
+        {
+          "$ref": "#/$defs/RawConstant"
+        }
+      ]
+    },
+    "FileConstant": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "FileConstant"
+        },
+        "value": {
+          "$ref": "#/$defs/File"
+        }
+      },
+      "required": [
+        "value"
+      ]
+    },
+    "RawConstant": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "RawConstant"
+        },
+        "value": {
+          "type": "number",
+          "$comment": "F64"
+        }
+      },
+      "required": [
+        "value"
+      ]
+    },
+    "ConstraintValue": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/UnitValue"
+        },
+        {
+          "$ref": "#/$defs/Datetime"
+        }
+      ]
+    },
+    "UnitValue": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "UnitValue"
+        },
+        "value": {
+          "type": "number",
+          "$comment": "F64"
+        },
+        "unit": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "value",
+        "unit"
+      ]
+    },
+    "Datetime": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Datetime"
+        },
+        "timestamp": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "timestamp"
+      ]
+    },
+    "Dimensionality": {
+      "type": "object",
+      "properties": {
+        "manifold": {
+          "type": "string",
+          "$comment": "I64"
+        },
+        "embedding": {
+          "type": "string",
+          "$comment": "I64"
+        }
+      },
+      "required": [
+        "manifold",
+        "embedding"
+      ]
+    },
+    "Region": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/RegionEquation"
+        }
+      ]
+    },
+    "RegionEquation": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "RegionEquation"
+        },
+        "equations": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "$comment": "Str"
+          }
+        }
+      },
+      "required": [
+        "equations"
+      ]
+    },
+    "MethodType": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/Barycenter"
+        },
+        {
+          "$ref": "#/$defs/Circumcenter"
+        }
+      ]
+    },
+    "Barycenter": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Barycenter"
+        }
+      },
+      "required": [
+      ]
+    },
+    "Circumcenter": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Circumcenter"
+        }
+      },
+      "required": [
+      ]
+    },
+    "PrimalDualRelation": {
+      "type": "object",
+      "properties": {
+        "primal": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "dual": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "method": {
+          "$ref": "#/$defs/MethodType"
+        }
+      },
+      "required": [
+        "primal",
+        "dual",
+        "method"
+      ]
+    },
+    "RelationType": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/Perimeter"
+        }
+      ]
+    },
+    "Perimeter": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Perimeter"
+        }
+      },
+      "required": [
+      ]
+    },
+    "SubmeshRelation": {
+      "type": "object",
+      "properties": {
+        "relation": {
+          "$ref": "#/$defs/RelationType"
+        },
+        "mesh": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "submesh": {
+          "type": "string",
+          "$comment": "Str"
+        }
+      },
+      "required": [
+        "relation",
+        "mesh",
+        "submesh"
+      ]
+    },
+    "Mesh": {
+      "type": "object",
+      "properties": {
+        "id": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "description": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "dimensionality": {
+          "$ref": "#/$defs/Dimensionality"
+        },
+        "vertex_count": {
+          "type": "string",
+          "$comment": "I64"
+        },
+        "edge_count": {
+          "type": "string",
+          "$comment": "I64"
+        },
+        "face_count": {
+          "type": "string",
+          "$comment": "I64"
+        },
+        "volume_count": {
+          "type": "string",
+          "$comment": "I64"
+        },
+        "regions": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Region"
+          }
+        },
+        "checksum": {
+          "type": "string",
+          "$comment": "Str"
+        },
+        "file": {
+          "$ref": "#/$defs/File"
+        }
+      },
+      "required": [
+        "id",
+        "description",
+        "dimensionality",
+        "vertex_count",
+        "edge_count",
+        "face_count",
+        "volume_count",
+        "regions",
+        "checksum",
+        "file"
+      ]
+    },
+    "DecapodeContext": {
+      "type": "object",
+      "properties": {
+        "constants": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/Constant"
+          }
+        },
+        "spatial_constraints": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/ConstraintValue"
+          }
+        },
+        "temporal_constraints": {
+          "type": "object",
+          "additionalProperties": {
+            "$ref": "#/$defs/ConstraintValue"
+          }
+        },
+        "primal_dual_relations": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/PrimalDualRelation"
+          }
+        },
+        "mesh_submesh_relations": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/SubmeshRelation"
+          }
+        },
+        "meshes": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Mesh"
+          }
+        }
+      },
+      "required": [
+        "constants",
+        "spatial_constraints",
+        "temporal_constraints",
+        "primal_dual_relations",
+        "mesh_submesh_relations",
+        "meshes"
+      ]
+    }
+  }
+}
