+++ decapodes/json/model_schema.json
@@ -1,7 +1,245 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
   "$defs": {
-    "SymDecapode": {
+    "Term": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/Var"
+        },
+        {
+          "$ref": "#/$defs/Lit"
+        },
+        {
+          "$ref": "#/$defs/AppCirc1"
+        },
+        {
+          "$ref": "#/$defs/App1"
+        },
+        {
+          "$ref": "#/$defs/App2"
+        },
+        {
+          "$ref": "#/$defs/Plus"
+        },
+        {
+          "$ref": "#/$defs/Mult"
+        },
+        {
+          "$ref": "#/$defs/Tan"
+        }
+      ]
+    },
+    "Var": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Var"
+        },
+        "name": {
+          "type": "string",
+          "$comment": "Sym"
+        }
+      },
+      "required": [
+        "name"
+      ]
+    },
+    "Lit": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Lit"
+        },
+        "name": {
+          "type": "string",
+          "$comment": "Sym"
+        }
+      },
+      "required": [
+        "name"
+      ]
+    },
+    "AppCirc1": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "AppCirc1"
+        },
+        "fs": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "$comment": "Sym"
+          }
+        },
+        "arg": {
+          "$ref": "#/$defs/Term"
+        }
+      },
+      "required": [
+        "fs",
+        "arg"
+      ]
+    },
+    "App1": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "App1"
+        },
+        "f": {
+          "type": "string",
+          "$comment": "Sym"
+        },
+        "arg": {
+          "$ref": "#/$defs/Term"
+        }
+      },
+      "required": [
+        "f",
+        "arg"
+      ]
+    },
+    "App2": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "App2"
+        },
+        "f": {
+          "type": "string",
+          "$comment": "Sym"
+        },
+        "arg1": {
+          "$ref": "#/$defs/Term"
+        },
+        "arg2": {
+          "$ref": "#/$defs/Term"
+        }
+      },
+      "required": [
+        "f",
+        "arg1",
+        "arg2"
+      ]
+    },
+    "Plus": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Plus"
+        },
+        "args": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Term"
+          }
+        }
+      },
+      "required": [
+        "args"
+      ]
+    },
+    "Mult": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Mult"
+        },
+        "args": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Term"
+          }
+        }
+      },
+      "required": [
+        "args"
+      ]
+    },
+    "Tan": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Tan"
+        },
+        "var": {
+          "$ref": "#/$defs/Term"
+        }
+      },
+      "required": [
+        "var"
+      ]
+    },
+    "Judgement": {
+      "type": "object",
+      "properties": {
+        "var": {
+          "type": "string",
+          "$comment": "Sym"
+        },
+        "dim": {
+          "type": "string",
+          "$comment": "Sym"
+        },
+        "space": {
+          "type": "string",
+          "$comment": "Sym"
+        }
+      },
+      "required": [
+        "var",
+        "dim",
+        "space"
+      ]
+    },
+    "Equation": {
+      "oneOf": [
+        {
+          "$ref": "#/$defs/Eq"
+        }
+      ]
+    },
+    "Eq": {
+      "type": "object",
+      "properties": {
+        "tag": {
+          "const": "Eq"
+        },
+        "lhs": {
+          "$ref": "#/$defs/Term"
+        },
+        "rhs": {
+          "$ref": "#/$defs/Term"
+        }
+      },
+      "required": [
+        "lhs",
+        "rhs"
+      ]
+    },
+    "DecaExpr": {
+      "type": "object",
+      "properties": {
+        "context": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Judgement"
+          }
+        },
+        "equations": {
+          "type": "array",
+          "items": {
+            "$ref": "#/$defs/Equation"
+          }
+        }
+      },
+      "required": [
+        "context",
+        "equations"
+      ]
+    },
+    "DecapodeSym": {
       "type": "object",
       "properties": {
         "type": "object",
@@ -140,7 +378,7 @@
         ]
       }
     },
-    "SymNamedDecapode": {
+    "NamedDecapodeSym": {
       "type": "object",
       "properties": {
         "type": "object",
@@ -284,7 +522,7 @@
         ]
       }
     },
-    "SymSummationDecapode": {
+    "SummationDecapodeSym": {
       "type": "object",
       "properties": {
         "type": "object",
