+++ decapodes/python/model.py
@@ -7,195 +7,6 @@ from intertypes import SafeInt, InterTypeBase
 from acsets import Ob, Hom, Attr, AttrType, Schema, ACSet
 
 
-class ModelHeader(InterTypeBase):
-    id: str | None
-    description: str
-    name: str
-    model_version: str
-    schema: str
-    schema_name: str
-
-
-class ContextHeader(InterTypeBase):
-    id: str | None
-    name: str
-    description: str
-    parent_model: str
-
-
-class ConfigurationHeader(InterTypeBase):
-    id: str | None
-    description: str
-    name: str
-    parent_context: str
-
-
-class File(InterTypeBase):
-    uri: str
-    format: str
-
-
-class LiteralParameter(InterTypeBase):
-    tag: Literal["LiteralParameter"] = Field(default="LiteralParameter", alias="_type", repr=False)
-    type: str
-    value: float
-
-
-class ConstantParameter(InterTypeBase):
-    tag: Literal["ConstantParameter"] = Field(default="ConstantParameter", alias="_type", repr=False)
-    type: str
-    value: str
-
-
-Parameter = Annotated[LiteralParameter | ConstantParameter, Field(discriminator="tag")]
-parameter_adapter: TypeAdapter[Parameter] = TypeAdapter(Parameter)
-
-
-class Condition(InterTypeBase):
-    type: str
-    value: str
-    domain_mesh: str
-
-
-class NetCDF(InterTypeBase):
-    tag: Literal["NetCDF"] = Field(default="NetCDF", alias="_type", repr=False)
-
-
-class CSV(InterTypeBase):
-    tag: Literal["CSV"] = Field(default="CSV", alias="_type", repr=False)
-
-
-DatasetFormat = Annotated[NetCDF | CSV, Field(discriminator="tag")]
-datasetformat_adapter: TypeAdapter[DatasetFormat] = TypeAdapter(DatasetFormat)
-
-
-class DatasetFile(InterTypeBase):
-    uri: str
-    format: "DatasetFormat"
-    shape: list[SafeInt]
-
-
-class Dataset(InterTypeBase):
-    type: str
-    name: str
-    description: str
-    file: "DatasetFile"
-
-
-class DecapodeConfiguration(InterTypeBase):
-    parameters: dict[str, "Parameter"]
-    initial_conditions: dict[str, "Condition"]
-    boundary_conditions: dict[str, "Condition"]
-    datasets: dict[str, "Dataset"]
-
-
-class ASKEMDecapodeConfiguration(InterTypeBase):
-    header: "ConfigurationHeader"
-    configuration: "DecapodeConfiguration"
-
-
-class FileConstant(InterTypeBase):
-    tag: Literal["FileConstant"] = Field(default="FileConstant", alias="_type", repr=False)
-    value: "File"
-
-
-class RawConstant(InterTypeBase):
-    tag: Literal["RawConstant"] = Field(default="RawConstant", alias="_type", repr=False)
-    value: float
-
-
-Constant = Annotated[FileConstant | RawConstant, Field(discriminator="tag")]
-constant_adapter: TypeAdapter[Constant] = TypeAdapter(Constant)
-
-
-class UnitValue(InterTypeBase):
-    tag: Literal["UnitValue"] = Field(default="UnitValue", alias="_type", repr=False)
-    value: float
-    unit: str
-
-
-class Datetime(InterTypeBase):
-    tag: Literal["Datetime"] = Field(default="Datetime", alias="_type", repr=False)
-    timestamp: str
-
-
-ConstraintValue = Annotated[UnitValue | Datetime, Field(discriminator="tag")]
-constraintvalue_adapter: TypeAdapter[ConstraintValue] = TypeAdapter(ConstraintValue)
-
-
-class Dimensionality(InterTypeBase):
-    manifold: SafeInt
-    embedding: SafeInt
-
-
-class RegionEquation(InterTypeBase):
-    tag: Literal["RegionEquation"] = Field(default="RegionEquation", alias="_type", repr=False)
-    equations: list[str]
-
-
-Region = Annotated[RegionEquation, Field(discriminator="tag")]
-region_adapter: TypeAdapter[Region] = TypeAdapter(Region)
-
-
-class Barycenter(InterTypeBase):
-    tag: Literal["Barycenter"] = Field(default="Barycenter", alias="_type", repr=False)
-
-
-class Circumcenter(InterTypeBase):
-    tag: Literal["Circumcenter"] = Field(default="Circumcenter", alias="_type", repr=False)
-
-
-MethodType = Annotated[Barycenter | Circumcenter, Field(discriminator="tag")]
-methodtype_adapter: TypeAdapter[MethodType] = TypeAdapter(MethodType)
-
-
-class PrimalDualRelation(InterTypeBase):
-    primal: str
-    dual: str
-    method: "MethodType"
-
-
-class Perimeter(InterTypeBase):
-    tag: Literal["Perimeter"] = Field(default="Perimeter", alias="_type", repr=False)
-
-
-RelationType = Annotated[Perimeter, Field(discriminator="tag")]
-relationtype_adapter: TypeAdapter[RelationType] = TypeAdapter(RelationType)
-
-
-class SubmeshRelation(InterTypeBase):
-    relation: "RelationType"
-    mesh: str
-    submesh: str
-
-
-class Mesh(InterTypeBase):
-    id: str
-    description: str
-    dimensionality: "Dimensionality"
-    vertex_count: SafeInt
-    edge_count: SafeInt
-    face_count: SafeInt
-    volume_count: SafeInt
-    regions: list["Region"]
-    checksum: str
-    file: "File"
-
-
-class DecapodeContext(InterTypeBase):
-    constants: dict[str, "Constant"]
-    spatial_constraints: dict[str, "ConstraintValue"]
-    temporal_constraints: dict[str, "ConstraintValue"]
-    primal_dual_relations: list["PrimalDualRelation"]
-    mesh_submesh_relations: list["SubmeshRelation"]
-    meshes: list["Mesh"]
-
-
-class ASKEMDecapodeContext(InterTypeBase):
-    header: "ContextHeader"
-    context: "DecapodeContext"
-
-
 class Var(InterTypeBase):
     tag: Literal["Var"] = Field(default="Var", alias="_type", repr=False)
     name: str
@@ -391,17 +202,5 @@ class SummationDecapodeSym(AbstractNamedDecapode):
     def read_json(cls, s: str):
         return super(SummationDecapodeSym, cls).read_json("SummationDecapodeSym", SchSummationDecapode.schema, s)
 
-
-
-class ASKEMDecapodeModel(InterTypeBase):
-    header: "ModelHeader"
-    model: "DecaExpr"
-
-
-class ASKEMDecapodeSimulationPlan(InterTypeBase):
-    header: "ModelHeader"
-    model: "DecaExpr"
-    context: "DecapodeContext"
-    configuration: "DecapodeConfiguration"
